separated KernelAttributes from ll_api
